fix: handle only one component/module per ts-file
