+++ projects/demo/src/modules/customization/dialogs/dialogs.component.ts
@@ -1,10 +1,11 @@
+import {default as example3Ts} from '!!raw-loader!./examples/prompt/prompt-options.ts';
 import {default as example1Ts} from '!!raw-loader!./examples/prompt/prompt.component.ts';
+import {default as example2Ts} from '!!raw-loader!./examples/prompt/prompt.service.ts';
 import {default as example1Less} from '!!raw-loader!./examples/prompt/prompt.style.less';
 import {default as example1Html} from '!!raw-loader!./examples/prompt/prompt.template.html';
 
 import {Component} from '@angular/core';
 import {changeDetection} from '../../../change-detection-strategy';
-import {FrontEndExample} from '../../interfaces/front-end-example';
 
 @Component({
     selector: 'dialogs',
@@ -12,9 +13,11 @@ import {FrontEndExample} from '../../interfaces/front-end-example';
     changeDetection,
 })
 export class DialogsComponent {
-    readonly example1: FrontEndExample = {
+    readonly example1 = {
         TypeScript: example1Ts,
         HTML: example1Html,
         LESS: example1Less,
+        'prompt.service.ts': example2Ts,
+        'prompt-options.ts': example3Ts,
     };
 }
