+++ projects/demo/src/modules/customization/dialogs/examples/prompt/prompt-options.ts
@@ -1,0 +1,4 @@
+export interface PromptOptions {
+    readonly heading: string;
+    readonly buttons: readonly [string, string];
+}
